+++ gl_util/src/model.cpp
@@ -189,7 +189,7 @@ unsigned int loadTexture(const std::string& imageFilepath) {
 
     glBindTexture(GL_TEXTURE_2D, texture);
     glTexImage2D(GL_TEXTURE_2D,
-                 0, format, width, height, 0,
+                 0, static_cast<GLint>(format), width, height, 0,
                  format, GL_UNSIGNED_BYTE, data);
     glGenerateMipmap(GL_TEXTURE_2D);
